Move classes to correct namespaces
