+++ Emby.Server.Implementations/ScheduledTasks/Triggers/IntervalTrigger.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using MediaBrowser.Model.Tasks;
 using Microsoft.Extensions.Logging;
 
-namespace Emby.Server.Implementations.ScheduledTasks
+namespace Emby.Server.Implementations.ScheduledTasks.Triggers
 {
     /// <summary>
     /// Represents a task trigger that runs repeatedly on an interval
