+++ Emby.Server.Implementations/ScheduledTasks/Triggers/StartupTrigger.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using MediaBrowser.Model.Tasks;
 using Microsoft.Extensions.Logging;
 
-namespace Emby.Server.Implementations.ScheduledTasks
+namespace Emby.Server.Implementations.ScheduledTasks.Triggers
 {
     /// <summary>
     /// Class StartupTaskTrigger
